backend/app/requirements.txt
aiopg==1.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2.13
django-annoying==0.10.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangochannelsrestframework==1.2.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Markdown==3.6
msgpack==1.0.8
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
redis==5.0.8
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
uuid==1.30
zope.interface==6.4.post2

requests==2.32.3
requests-oauthlib==2.0.0
oauthlib==3.2.2